requirements.txt
Django~=4.0.2
djangorestframework~=3.13.1
Pillow~=9.0.1
jwt~=1.3.1